Do not include <windows.h> in <errors.hpp>.
Including the windows main header interfers with programmes using
pathie that include it on their own. Since it was only needed for the
definitions of DWORD and HRESULT, those types can simply be spelled
out and it still compiles, removing the need for windows.h in the
header file.

windows.h is now only included inside pathie's .cpp source files,
which does not cause the aforementioned problem.