Change license from GPL-3 to BSD-2-clause.
This library is on the first page of the Google results if you search
for MinGW and C++ Unicode filenames. Still, appearently nobody uses
it. I suppose this is due to the GPL license, so I decided to
relicense the entire thing under BSD-2-clause. This should do away
with any license problems people saw in the library.