Handle ENOTDIR errno as indicating a path does not exist.
Previously, Path::exists() would produce an exception when a parent of
the checked path existed, but was not a directory. For example, if
the user called `Path("foo/baz.txt").exists()` when "foo" was a file.

This change makes Pathie act more similarly to path libraries such as
Python's pathlib, which returns false given the same inputs.